pyproject.toml
[project]
name = "hanjie-riddle"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.1.1",
]
